# 8352972: PPC64: Intrinsify Unsafe::setMemory